cop_dag/aggr_fn, coprocessor/dag/aggr_fn: extract `aggr_fn`
Move aggr_fn mod from `tikv` into `cop_dag`
